chore(cr): move tests into
